Move SDK code into sdk directory.
