+++ web_transport/sdk/impl/tests/quic_transport_owt_end_to_end_test.cc
@@ -25,7 +25,7 @@
 #include "owt/quic/quic_transport_client_interface.h"
 #include "owt/quic/quic_transport_factory.h"
 #include "owt/quic/quic_transport_stream_interface.h"
-#include "owt/web_transport/impl/quic_transport_owt_client_impl.h"
+#include "owt/web_transport/sdk/impl/quic_transport_owt_client_impl.h"
 #include "testing/gmock/include/gmock/gmock.h"
 #include "testing/gtest/include/gtest/gtest.h"
