+++ web_transport/sdk/impl/version_info_values.h.template
