+++ web_transport/sdk/resources/ssl/certificates/proof_source_pkcs12_test.pfx
